tourn: list players
